adding integration tests
